중복 체크와 임시 멤버 삭제 회원 가입시 검증 추가

+++ src/main/java/com/example/capstone/member/service/MemberServiceImpl.java
@@ -90,7 +90,14 @@ public class MemberServiceImpl implements MemberService {
 
         log.info("deleteTempMember service start............");
 
-        memberRepository.deleteById(id);
+        Member tempMember = memberRepository.findById(id).orElseThrow(() -> new ExceptionHandler(ErrorStatus.MEMBER_NOT_FOUND));
+
+        if (tempMember.getPassword() == null) {
+            memberRepository.deleteById(id);
+        } else {
+            throw new ExceptionHandler(ErrorStatus.TEMP_MEMBER_NOT_FOUND);
+        }
+
 
     }
